tests/wpt/web-platform-tests/webxr/xrSession_requestAnimationFrame_data_valid.https.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src="resources/webxr_util.js"></script>
  <canvas></canvas>

  <script>
    const testName = "RequestAnimationFrame resolves with good data";

    const identityMatrix = new Float32Array([
      1, 0, 0, 0,
      0, 1, 0, 0,
      0, 0, 1, 0,
      0, 0, 0, 1
    ]);

    const rightFakeXRViewInit = {
      eye:"right",
      projectionMatrix: identityMatrix,
      viewMatrix: identityMatrix
    };

    const leftFakeXRViewInit = {
      eye:"left",
      projectionMatrix: identityMatrix,
      viewMatrix: identityMatrix
    };

    const fakeDeviceInitOptions = { supportsImmersive:true };
    const sessionOptions = { mode: 'immersive-vr' };

    let testSession;

    let testFunction = function(session, testDeviceController) {
      testSession = session;
      return session.requestReferenceSpace({ type: 'stationary', subtype: 'eye-level' })
        .then((referenceSpace) => new Promise((resolve) => {

          function onFrame(time, xrFrame) {
            assert_true(xrFrame instanceof XRFrame);

            let viewerPose = xrFrame.getViewerPose(referenceSpace);

            assert_not_equals(viewerPose, null);
            for(let i = 0; i < identityMatrix.length; i++) {
              assert_equals(viewerPose.poseModelMatrix[i], identityMatrix[i]);
            }

            assert_not_equals(viewerPose.views, null);
            assert_equals(viewerPose.views.length, 2);

            assert_not_equals(viewerPose.views[0].viewMatrix, null);
            assert_equals(viewerPose.views[0].viewMatrix.length, 16);
            assert_not_equals(viewerPose.views[1].viewMatrix, null);
            assert_equals(viewerPose.views[1].viewMatrix.length, 16);

            // Test does not complete until the returned promise resolves.
            resolve();
          }

          testDeviceController.setXRPresentationFrameData(
            identityMatrix,
            [rightFakeXRViewInit, leftFakeXRViewInit]
          );

          testSession.requestAnimationFrame(onFrame);
        })
      );
    }

    xr_session_promise_test(
      testName, testFunction, fakeDeviceInitOptions, sessionOptions);
  </script>
</body>
